olorin-server/test/contracts/test_polling_api_contract.py
"""
Contract Test: Polling API
Feature: 005-polling-and-persistence

Tests ALL 4 polling endpoints against the OpenAPI specification using contract testing.
Tests are expected to FAIL initially (endpoints not implemented yet).

Adaptive Polling Strategy:
- Fast (500ms): status=IN_PROGRESS AND lifecycle_stage=IN_PROGRESS
- Normal (2s): status=SETTINGS AND lifecycle_stage=SETTINGS
- Slow (5s): status=COMPLETED OR lifecycle_stage=COMPLETED

SYSTEM MANDATE Compliance:
- Contract-first approach (test before implementation)
- No mock data (tests validate real API behavior)
- All endpoints tested against OpenAPI schema
- All error codes validated (401, 403, 404, 429, 500)
- ETag conditional GET behavior tested
- Rate limiting tested with Retry-After header

Endpoints Tested:
1. GET /polling/investigation-state/{id} (poll with ETag support)
2. GET /polling/investigation-state/{id}/changes (delta polling)
3. GET /polling/active-investigations (list polling)
4. GET /polling/health (health check, no auth)
"""

import pytest
from pathlib import Path
import yaml
from typing import Dict, Any
from openapi_core import OpenAPI


SPEC_PATH = Path(__file__).parent.parent.parent.parent / "specs" / "005-polling-and-persistence" / "contracts" / "polling-api.yaml"


@pytest.fixture(scope="module")
def openapi_spec() -> Dict[str, Any]:
    """Load Polling API OpenAPI specification from YAML file."""
    with open(SPEC_PATH, "r") as f:
        return yaml.safe_load(f)


@pytest.fixture(scope="module")
def openapi_validator(openapi_spec: Dict[str, Any]):
    """Create OpenAPI validator from specification."""
    return OpenAPI.from_dict(openapi_spec)


class TestPollingAPIContract:
    """Contract tests for Polling API endpoints."""

    @pytest.mark.unit
    def test_poll_investigation_state_schema(self, openapi_validator):
        """
        Test GET /polling/investigation-state/{id} request/response schema.

        Validates:
        - Path parameter validation (investigation_id pattern)
        - Response schema matches PollingStateResponse
        - ETag header present in response
        - X-Recommended-Interval header with valid interval

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_investigation_state_conditional_get(self, openapi_validator):
        """
        Test GET /polling/investigation-state/{id} with If-None-Match header.

        Validates:
        - If-None-Match header support
        - 304 Not Modified response when ETag matches
        - X-Recommended-Interval header in 304 response
        - ETag comparison logic

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_investigation_state_adaptive_intervals(
        self,
        openapi_validator
    ):
        """
        Test adaptive polling interval recommendations.

        Validates recommended_interval_ms based on state:
        - Fast (500ms): status=IN_PROGRESS AND lifecycle_stage=IN_PROGRESS
        - Normal (2s): status=SETTINGS AND lifecycle_stage=SETTINGS
        - Slow (5s): status=COMPLETED OR lifecycle_stage=COMPLETED

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_investigation_state_client_interval_header(
        self,
        openapi_validator
    ):
        """
        Test X-Polling-Interval request header validation.

        Validates:
        - Header accepts integer values
        - Minimum value 500ms
        - Maximum value 30000ms (30s)
        - Used for server analytics only

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_state_changes_schema(self, openapi_validator):
        """
        Test GET /polling/investigation-state/{id}/changes schema.

        Validates:
        - since_version query parameter required
        - include_snapshot query parameter optional
        - Response schema matches StateChangesResponse
        - changes array contains StateChange objects
        - current_snapshot object structure

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_state_changes_delta_response(self, openapi_validator):
        """
        Test delta polling response format.

        Validates:
        - from_version matches request parameter
        - to_version is current version
        - changes array ordered by version
        - change_type enum values
        - fields_changed array contains field paths
        - details object contains change-specific data

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_state_changes_no_changes(self, openapi_validator):
        """
        Test 304 Not Modified response when no changes.

        Validates:
        - 304 status when since_version equals current version
        - ETag header matches since_version
        - No response body

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_active_investigations_schema(self, openapi_validator):
        """
        Test GET /polling/active-investigations schema validation.

        Validates:
        - Query parameters (status, lifecycle_stage, limit, offset)
        - Response schema matches ActiveInvestigationsResponse
        - investigations array contains InvestigationSummary objects
        - Pagination metadata (total, limit, offset)
        - timestamp field present

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_active_investigations_filtering(self, openapi_validator):
        """
        Test filtering parameters for active investigations.

        Validates:
        - status filter enum values
        - lifecycle_stage filter enum values
        - Multiple filters can be combined
        - Results match filter criteria

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_active_investigations_pagination(self, openapi_validator):
        """
        Test pagination parameters.

        Validates:
        - limit parameter (min 1, max 100, default 20)
        - offset parameter (min 0, default 0)
        - total count in response
        - investigations array size <= limit

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_poll_active_investigations_etag(self, openapi_validator):
        """
        Test ETag support for active investigations list.

        Validates:
        - ETag header in 200 response
        - If-None-Match header support
        - 304 Not Modified when list unchanged
        - ETag format (e.g., "list-version-42")

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_polling_health_check_schema(self, openapi_validator):
        """
        Test GET /polling/health schema validation.

        Validates:
        - No authentication required (public endpoint)
        - Response schema matches PollingHealthResponse
        - status enum values (healthy, degraded, unavailable)
        - server_time timestamp present
        - recommended_intervals object structure

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_polling_health_check_status_codes(self, openapi_validator):
        """
        Test health check status codes.

        Validates:
        - 200 OK for healthy or degraded status
        - 503 Service Unavailable for unavailable status
        - Response body present in both cases

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_polling_health_check_intervals_by_load(self, openapi_validator):
        """
        Test recommended intervals adjust based on server load.

        Validates:
        - healthy status: fast=500ms, normal=2000ms, slow=5000ms
        - degraded status: intervals increased (e.g., fast=1000ms)
        - server_load object contains connection metrics
        - load_percentage correlates with interval adjustments

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_rate_limiting_response(self, openapi_validator):
        """
        Test 429 Too Many Requests response for all polling endpoints.

        Validates:
        - 429 status code
        - Retry-After header present (seconds to wait)
        - X-RateLimit-Limit header (max requests per window)
        - X-RateLimit-Remaining header (remaining requests)
        - X-RateLimit-Reset header (unix timestamp)
        - Response body matches ErrorResponse schema

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_error_response_schemas(self, openapi_validator):
        """
        Test all error response schemas match OpenAPI specification.

        Validates error responses for:
        - 400 Bad Request (invalid query parameters)
        - 401 Unauthorized (missing auth)
        - 403 Forbidden (insufficient permissions)
        - 404 Not Found (investigation not found, version too old)
        - 429 Too Many Requests (rate limit exceeded)
        - 500 Server Error (internal error)

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")

    @pytest.mark.unit
    def test_polling_response_minimal_payload(self, openapi_validator):
        """
        Test polling responses use minimal payloads.

        Validates:
        - PollingStateResponse excludes full settings/progress/results
        - Only includes summary data (percent_complete, risk_score)
        - Optimized for high-frequency polling bandwidth

        Expected to FAIL initially (endpoint not implemented).
        """
        pytest.skip("Endpoint not implemented yet - will implement in T016")
